+++ 2projekt/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks.Sources;
+using _2projekt;
 
 internal class Program
 {
@@ -11,38 +12,7 @@ internal class Program
             {
                 Console.WriteLine("podaj rownanie ktorego chcesz znac wynik:");
                 var input = Console.ReadLine();
-                var lista = input.Split("+").ToList();
-
-                var iloscLiczbDodatnich = lista.Count();
-                var liczby = new List<Int64>();
-
-                foreach (var item in lista.ToList())
-                {
-                   if (item.Length > 1)
-                    {
-                        lista.Remove(item);
-                        lista.AddRange(item.Split("-"));
-                    }
-                }
-                Int64 wynik = 0;
-                foreach (var item in lista)
-                {
-                  Int64.TryParse(item, out var liczba);
-                    liczby.Add(liczba);
-                }
-
-                for (int i = 0; i < liczby.Count(); i++)
-                {
-
-                    if (i < iloscLiczbDodatnich)
-                    {
-                        wynik = Calculator.Add(wynik, liczby[i]);
-                    }
-                    else
-                    {
-                        wynik = Calculator.Minus(wynik, liczby[i]);
-                    }
-                }
+                var wynik = Calculator.Process(input);
 
                 Console.WriteLine($"wynik to: {wynik}");
             }
@@ -55,23 +25,3 @@ internal class Program
     }
 }
 
-public static class Calculator
-
-{
-    public static Int64 Multiply(Int64 x, Int64 y)
-    {
-        return x * y;
-    }
-    public static Int64 Divide(Int64 x, Int64 y)
-    {
-        return x / y;
-    }
-    public static Int64 Minus(Int64 x, Int64 y)
-    {
-        return x - y;
-    }
-    public static Int64 Add(Int64 x, Int64 y)
-    {
-        return x + y;
-    }
-}
